refactor: couting(집계) 도메인을 audit으로 이동
- 패키지 순환참조 해결을 위함
