+++ src/main/java/com/gonghak98/v2/audit/domain/counting/AreaCreditSummary.java
@@ -1,4 +1,4 @@
-package com.gonghak98.v2.report.domain.counting;
+package com.gonghak98.v2.audit.domain.counting;
 
 import com.gonghak98.v2.audit.domain.constant.AbeekType;
 import com.gonghak98.v2.audit.domain.dto.AuditCompletedCourse;
