+++ src/main/java/com/gonghak98/v2/audit/domain/counting/CreditCountResult.java
@@ -1,4 +1,4 @@
-package com.gonghak98.v2.report.domain.counting;
+package com.gonghak98.v2.audit.domain.counting;
 
 public record CreditCountResult(double completedCredits,
                                 double requiredCredits) {
