+++ src/main/java/com/gonghak98/v2/report/controller/dto/ReportResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.gonghak98.v2.audit.domain.constant.AbeekType;
 import com.gonghak98.v2.audit.domain.dto.AuditCompletedCourse;
 import com.gonghak98.v2.audit.domain.dto.NonPassResult;
-import com.gonghak98.v2.report.domain.counting.AreaCreditSummary;
+import com.gonghak98.v2.audit.domain.counting.AreaCreditSummary;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
